refactor: refresh control
